Report mysql status to "service --status-all" properly.
Two changes are required to get this working:

First, System V's "service" command just runs "/etc/init.d/mysql status".  But that will incorrectly report that mysql is not running if not run with elevated privileges.   This is because /etc/mysql/debian.cnf is not other-readable; removing the reference to this file from the call to mysqladmin seems to work correctly without requiring sudo.

Second, the "service" command searches the source of /etc/init.d/mysql for "\Wstatus)" to decide whether it can determine service status.  Removing the single quotes from the case statement allows this to get past that condition.